requirements.txt
# Flowmap_creator
numpy
pillow